brands get url from uframe, episodes get url from data
